Move classes nested within BenchmarkModelComponent to the top level.
RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=187628039
